analyzer/requirements.txt
transformers==4.37.2
torch>=2.0,<3
kafka-python==2.0.2